+++ src/grid/mod.rs
@@ -1,6 +1,7 @@
+pub mod grid;
 pub mod message;
 pub mod cell;
-pub mod grid;
 pub mod traverse;
 pub mod bit;
 pub mod util;
+pub mod image;
